Reserve storage-bound items across auto-fill and auto-collect
Both auto-fill (when it would dump current items to the core in
order to fetch the next request) and auto-collect (when it would
empty the drone into the core after a successful pickup) used to
ignore storage-fill entirely. So a copper run picked up at a drill
and tagged for a configured vault was just shoveled back into the
core, defeating the whole feature.

Promoted isItemReservedForStorage from a private auto-fill helper
to storageFill.isItemReservedForStorage(item, team) and taught it
about reachability:

- with autopilot on, any team storage that needs the item counts as
  reachable (the drone will route)
- with autopilot off, only storages within the player's
  Vars.buildingRange count, because otherwise the player can't
  manually carry the item there anyway

Both auto-fill's core dump path and auto-collect's "drone-full near
core" dump now consult this helper and bail early when the item is
reserved. auto-fill's old inline helper is gone -- both paths share
the storage-fill version.

Bumped to v1.00.
