+++ mod.json
@@ -3,7 +3,7 @@
     "displayName": "[orange]extended[gold] UI",
     "author": "wizer",
     "description": "Extend your ui with some useful things.\n\nYou can easily disable any of the mod features in the settings -> extended UI.\n\nIf u have any bugs create an issue on github or contact me in discord: wizer1233\n\nTranslators:\n1stvaliduser - UA\nCoCo Snow - DE\n\nContributors:\nRushieWashie, Garen7, 7003Mars (Mars)",
-    "version": "0.99",
+    "version": "1.00",
     "minGameVersion": "136",
     "hidden": true
 }
