+++ scripts/interact/auto-fill.js
@@ -1,6 +1,5 @@
 const timer = require("extended-ui/interact/interact-timer");
 const coreLimits = require("extended-ui/interact/core-limits");
-const storageConfig = require("extended-ui/interact/storage-config");
 const storageFill = require("extended-ui/interact/storage-fill");
 const turretConfig = require("extended-ui/interact/turret-config");
 
@@ -68,7 +67,7 @@ Events.run(Trigger.update, () => {
     if (!isCoreAvailible || !request || !player.within(core, Vars.buildingRange)) return;
 
     if (stack.amount) {
-        if (isItemReservedForStorage(stack.item, team)) return;
+        if (storageFill.isItemReservedForStorage(stack.item, team)) return;
         Call.transferInventory(player, core);
         if (stack.amount > 0) {
             Call.dropItem(0);
@@ -78,22 +77,6 @@ Events.run(Trigger.update, () => {
     }
     timer.increase();
 });
-
-function isItemReservedForStorage(item, team) {
-    if (!Core.settings.getBool("eui-storage-fill", false)) return false;
-    const data = team.data();
-    if (!data || !data.buildings) return false;
-    let reserved = false;
-    data.buildings.each(b => {
-        if (reserved) return;
-        try {
-            if (!storageFill.isManagedStorage(b.block)) return;
-            const threshold = storageConfig.getThreshold(b, item);
-            if (threshold > 0 && b.items && b.items.get(item) < threshold) reserved = true;
-        } catch (e) {}
-    });
-    return reserved;
-}
 
 function getBestAmmo(turret, core) {
     let best = null;
